requirements.txt
fastapi
uvicorn
streamlit
chromadb==0.5.0
sentence-transformers
python-dotenv
google-generativeai
selenium
requests
pydantic
typing_extensions
